Move getColorEnumValues from utils.ts into model/game/card
