--- frontend/src/utils.ts
@@ -1,5 +1,0 @@
-import { Color } from '@/model/game/card';
-
-export function getColorEnumValues(): Color[] {
-    return Object.keys(Color).map(key => Color[key]);
-}
